requirements.txt
mimesis==4.1.3
numpy==1.19.5
pandas==1.1.5
py4j==0.10.9
pyspark==3.1.2
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
